pyproject.toml
[project]
name = "deepgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.2",
    "grpcio>=1.75.0",
    "llama-index-core>=0.14.2",
    "nebula3-python>=3.8.3",
    "numpy>=2.3.3",
    "ollama>=0.5.4",
    "openai>=1.108.0",
    "orjson>=3.11.3",
    "polars>=1.33.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "tavily-python>=0.7.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
